change TcpChannelSupplier from class to interface
